refactor: Invoice list component
update the list body
create customs components
create a custom hook
